12-15

+++ src/main/java/com/app/japub/controller/CommentController.java
@@ -22,6 +22,7 @@ import com.app.japub.domain.dto.CommentsDto;
 import com.app.japub.domain.dto.Criteria;
 import com.app.japub.domain.service.admin.AdminService;
 import com.app.japub.domain.service.comment.CommentService;
+import com.app.japub.domain.service.user.UserService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -31,35 +32,49 @@ import lombok.RequiredArgsConstructor;
 public class CommentController {
 	private final CommentService commentService;
 	private final AdminService adminService;
+	private final UserService userService;
 	private final HttpSession session;
 	private static final ResponseEntity<Void> SUCCESS_CODE = ResponseEntity.ok().build();
 	private static final ResponseEntity<Void> ERROR_CODE = ResponseEntity.badRequest().build();
 
-	@PostMapping(value = "/{boardNum}", consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Void> insert(@RequestBody CommentDto commentDto, @PathVariable Long boardNum) {
+	@PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<Void> insert(@RequestBody CommentDto commentDto, Long boardNum) {
 		Long userNum = (Long) session.getAttribute(SessionUtil.KEY);
+
 		if (userNum == null) {
 			return ERROR_CODE;
 		}
+
+		if (userService.findByUserNum(userNum) == null) {
+			return ResponseEntity.notFound().build();
+		}
+
 		commentDto.setUserNum(userNum);
 		commentDto.setBoardNum(boardNum);
-		if (!commentService.insert(commentDto)) {
-			return ERROR_CODE;
-		}
-		return SUCCESS_CODE;
+
+		return commentService.insert(commentDto) ? SUCCESS_CODE : ERROR_CODE;
 	}
 
 	@PatchMapping(value = "/{commentNum}", consumes = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Void> update(@RequestBody CommentDto commentDto, @PathVariable Long commentNum) {
+
 		Long userNum = (Long) session.getAttribute(SessionUtil.KEY);
+
 		if (userNum == null) {
 			return ERROR_CODE;
 		}
+
+		if (userService.findByUserNum(userNum) == null) {
+			return ResponseEntity.notFound().build();
+		}
+
 		commentDto.setUserNum(userNum);
 		commentDto.setCommentNum(commentNum);
-		boolean result = SessionUtil.isAdmin(session) ? adminService.updateComment(commentDto)
+
+		boolean isUpdated = SessionUtil.isAdmin(session) ? adminService.updateComment(commentDto)
 				: commentService.update(commentDto);
-		return result ? SUCCESS_CODE : ERROR_CODE;
+
+		return isUpdated ? SUCCESS_CODE : ERROR_CODE;
 	}
 
 	@DeleteMapping(value = "/{commentNum}")
@@ -70,19 +85,22 @@ public class CommentController {
 			return ERROR_CODE;
 		}
 
-		boolean result = SessionUtil.isAdmin(session) ? adminService.deleteByCommentNum(commentNum)
+		if (userService.findByUserNum(userNum) == null) {
+			return ResponseEntity.notFound().build();
+		}
+
+		boolean isDeleted = SessionUtil.isAdmin(session) ? adminService.deleteByCommentNum(commentNum)
 				: commentService.delete(userNum, commentNum);
 
-		return result ? SUCCESS_CODE : ERROR_CODE;
+		return isDeleted ? SUCCESS_CODE : ERROR_CODE;
 	}
 
-	@GetMapping(value = "/{boardNum}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<CommentsDto> list(@PathVariable Long boardNum, Criteria criteria) {
+	@GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<CommentsDto> list(Long boardNum, Criteria criteria) {
 		List<CommentDto> comments = commentService.findByCriteriaAndBoardNum(criteria, boardNum);
-		criteria.setPage(criteria.getPage() + 1);
-		int nextCountPage = commentService.getNextPageCount(criteria, boardNum);
-		CommentsDto commentsDto = new CommentsDto(comments, nextCountPage);
-		return new ResponseEntity<CommentsDto>(commentsDto, HttpStatus.OK);
+		int nextPageCount = commentService.getNextPageCount(criteria, boardNum);
+
+		return new ResponseEntity<CommentsDto>(new CommentsDto(comments, nextPageCount), HttpStatus.OK);
 	}
 
 }
